Refactor meta_tags
